+++ cvxpy/tests/test_atoms.py
@@ -18,6 +18,7 @@ along with CVXPY.  If not, see <http://www.gnu.org/licenses/>.
 """
 
 import cvxpy
+import cvxpy.settings as s
 from cvxpy.atoms import *
 from cvxpy.expressions.variables import Variable, NonNegative, Bool, Int
 from cvxpy.expressions.constants import Parameter
@@ -63,11 +64,11 @@ class TestAtoms(BaseTest):
         atom = normInf(exp)
         # self.assertEquals(atom.name(), "normInf(x + y)")
         self.assertEquals(atom.size, (1,1))
-        self.assertEquals(atom.curvature, u.Curvature.CONVEX_KEY)
+        self.assertEquals(atom.curvature, s.CONVEX)
         assert atom.is_convex()
         assert (-atom).is_concave()
-        self.assertEquals(normInf(atom).curvature, u.Curvature.CONVEX_KEY)
-        self.assertEquals(normInf(-atom).curvature, u.Curvature.CONVEX_KEY)
+        self.assertEquals(normInf(atom).curvature, s.CONVEX)
+        self.assertEquals(normInf(-atom).curvature, s.CONVEX)
 
     # Test the norm1 class.
     def test_norm1(self):
@@ -75,9 +76,9 @@ class TestAtoms(BaseTest):
         atom = norm1(exp)
         # self.assertEquals(atom.name(), "norm1(x + y)")
         self.assertEquals(atom.size, (1,1))
-        self.assertEquals(atom.curvature, u.Curvature.CONVEX_KEY)
-        self.assertEquals(norm1(atom).curvature, u.Curvature.CONVEX_KEY)
-        self.assertEquals(norm1(-atom).curvature, u.Curvature.CONVEX_KEY)
+        self.assertEquals(atom.curvature, s.CONVEX)
+        self.assertEquals(norm1(atom).curvature, s.CONVEX)
+        self.assertEquals(norm1(-atom).curvature, s.CONVEX)
 
     # Test the norm2 class.
     def test_norm2(self):
@@ -85,9 +86,9 @@ class TestAtoms(BaseTest):
         atom = norm2(exp)
         # self.assertEquals(atom.name(), "norm2(x + y)")
         self.assertEquals(atom.size, (1, 1))
-        self.assertEquals(atom.curvature, u.Curvature.CONVEX_KEY)
-        self.assertEquals(norm2(atom).curvature, u.Curvature.CONVEX_KEY)
-        self.assertEquals(norm2(-atom).curvature, u.Curvature.CONVEX_KEY)
+        self.assertEquals(atom.curvature, s.CONVEX)
+        self.assertEquals(norm2(atom).curvature, s.CONVEX)
+        self.assertEquals(norm2(-atom).curvature, s.CONVEX)
 
     # Test the power class.
     def test_power(self):
@@ -104,16 +105,16 @@ class TestAtoms(BaseTest):
                 self.assertEquals(atom.size, size)
 
                 if p > 1 or p < 0:
-                    self.assertEquals(atom.curvature, u.Curvature.CONVEX_KEY)
+                    self.assertEquals(atom.curvature, s.CONVEX)
                 elif p == 1:
-                    self.assertEquals(atom.curvature, u.Curvature.AFFINE_KEY)
+                    self.assertEquals(atom.curvature, s.AFFINE)
                 elif p == 0:
-                    self.assertEquals(atom.curvature, u.Curvature.CONSTANT_KEY)
+                    self.assertEquals(atom.curvature, s.CONSTANT)
                 else:
-                    self.assertEquals(atom.curvature, u.Curvature.CONCAVE_KEY)
+                    self.assertEquals(atom.curvature, s.CONCAVE)
 
                 if p != 1:
-                    self.assertEquals(atom.sign, u.Sign.POSITIVE_KEY)
+                    self.assertEquals(atom.sign, s.POSITIVE)
 
                 # Test copy with args=None
                 copy = atom.copy()
@@ -134,8 +135,8 @@ class TestAtoms(BaseTest):
     def test_geo_mean(self):
         atom = geo_mean(self.x)
         self.assertEquals(atom.size, (1, 1))
-        self.assertEquals(atom.curvature, u.Curvature.CONCAVE_KEY)
-        self.assertEquals(atom.sign, u.Sign.POSITIVE_KEY)
+        self.assertEquals(atom.curvature, s.CONCAVE)
+        self.assertEquals(atom.sign, s.POSITIVE)
         # Test copy with args=None
         copy = atom.copy()
         self.assertTrue(type(copy) is type(atom))
@@ -155,65 +156,65 @@ class TestAtoms(BaseTest):
     def test_harmonic_mean(self):
         atom = harmonic_mean(self.x)
         self.assertEquals(atom.size, (1, 1))
-        self.assertEquals(atom.curvature, u.Curvature.CONCAVE_KEY)
-        self.assertEquals(atom.sign, u.Sign.POSITIVE_KEY)
+        self.assertEquals(atom.curvature, s.CONCAVE)
+        self.assertEquals(atom.sign, s.POSITIVE)
 
     # Test the pnorm class.
     def test_pnorm(self):
         atom = pnorm(self.x, p=1.5)
         self.assertEquals(atom.size, (1, 1))
-        self.assertEquals(atom.curvature, u.Curvature.CONVEX_KEY)
-        self.assertEquals(atom.sign, u.Sign.POSITIVE_KEY)
+        self.assertEquals(atom.curvature, s.CONVEX)
+        self.assertEquals(atom.sign, s.POSITIVE)
 
         atom = pnorm(self.x, p=1)
         self.assertEquals(atom.size, (1, 1))
-        self.assertEquals(atom.curvature, u.Curvature.CONVEX_KEY)
-        self.assertEquals(atom.sign, u.Sign.POSITIVE_KEY)
+        self.assertEquals(atom.curvature, s.CONVEX)
+        self.assertEquals(atom.sign, s.POSITIVE)
 
         atom = pnorm(self.x, p=2)
         self.assertEquals(atom.size, (1, 1))
-        self.assertEquals(atom.curvature, u.Curvature.CONVEX_KEY)
-        self.assertEquals(atom.sign, u.Sign.POSITIVE_KEY)
+        self.assertEquals(atom.curvature, s.CONVEX)
+        self.assertEquals(atom.sign, s.POSITIVE)
 
         atom = pnorm(self.x, p='inf')
         self.assertEquals(atom.size, (1, 1))
-        self.assertEquals(atom.curvature, u.Curvature.CONVEX_KEY)
-        self.assertEquals(atom.sign, u.Sign.POSITIVE_KEY)
+        self.assertEquals(atom.curvature, s.CONVEX)
+        self.assertEquals(atom.sign, s.POSITIVE)
 
         atom = pnorm(self.x, p='Inf')
         self.assertEquals(atom.size, (1, 1))
-        self.assertEquals(atom.curvature, u.Curvature.CONVEX_KEY)
-        self.assertEquals(atom.sign, u.Sign.POSITIVE_KEY)
+        self.assertEquals(atom.curvature, s.CONVEX)
+        self.assertEquals(atom.sign, s.POSITIVE)
 
         atom = pnorm(self.x, p=np.inf)
         self.assertEquals(atom.size, (1, 1))
-        self.assertEquals(atom.curvature, u.Curvature.CONVEX_KEY)
-        self.assertEquals(atom.sign, u.Sign.POSITIVE_KEY)
+        self.assertEquals(atom.curvature, s.CONVEX)
+        self.assertEquals(atom.sign, s.POSITIVE)
 
         atom = pnorm(self.x, p=.5)
         self.assertEquals(atom.size, (1, 1))
-        self.assertEquals(atom.curvature, u.Curvature.CONCAVE_KEY)
-        self.assertEquals(atom.sign, u.Sign.POSITIVE_KEY)
+        self.assertEquals(atom.curvature, s.CONCAVE)
+        self.assertEquals(atom.sign, s.POSITIVE)
 
         atom = pnorm(self.x, p=.7)
         self.assertEquals(atom.size, (1, 1))
-        self.assertEquals(atom.curvature, u.Curvature.CONCAVE_KEY)
-        self.assertEquals(atom.sign, u.Sign.POSITIVE_KEY)
+        self.assertEquals(atom.curvature, s.CONCAVE)
+        self.assertEquals(atom.sign, s.POSITIVE)
 
         atom = pnorm(self.x, p=-.1)
         self.assertEquals(atom.size, (1, 1))
-        self.assertEquals(atom.curvature, u.Curvature.CONCAVE_KEY)
-        self.assertEquals(atom.sign, u.Sign.POSITIVE_KEY)
+        self.assertEquals(atom.curvature, s.CONCAVE)
+        self.assertEquals(atom.sign, s.POSITIVE)
 
         atom = pnorm(self.x, p=-1)
         self.assertEquals(atom.size, (1, 1))
-        self.assertEquals(atom.curvature, u.Curvature.CONCAVE_KEY)
-        self.assertEquals(atom.sign, u.Sign.POSITIVE_KEY)
+        self.assertEquals(atom.curvature, s.CONCAVE)
+        self.assertEquals(atom.sign, s.POSITIVE)
 
         atom = pnorm(self.x, p=-1.3)
         self.assertEquals(atom.size, (1, 1))
-        self.assertEquals(atom.curvature, u.Curvature.CONCAVE_KEY)
-        self.assertEquals(atom.sign, u.Sign.POSITIVE_KEY)
+        self.assertEquals(atom.curvature, s.CONCAVE)
+        self.assertEquals(atom.sign, s.POSITIVE)
 
         # Test copy with args=None
         copy = atom.copy()
@@ -233,11 +234,11 @@ class TestAtoms(BaseTest):
     def test_quad_over_lin(self):
         # Test quad_over_lin DCP.
         atom = quad_over_lin(square(self.x), self.a)
-        self.assertEquals(atom.curvature, u.Curvature.CONVEX_KEY)
+        self.assertEquals(atom.curvature, s.CONVEX)
         atom = quad_over_lin(-square(self.x), self.a)
-        self.assertEquals(atom.curvature, u.Curvature.CONVEX_KEY)
+        self.assertEquals(atom.curvature, s.CONVEX)
         atom = quad_over_lin(sqrt(self.x), self.a)
-        self.assertEquals(atom.curvature, u.Curvature.UNKNOWN_KEY)
+        self.assertEquals(atom.curvature, s.UNKNOWN)
         assert not atom.is_dcp()
 
         # Test quad_over_lin size validation.
@@ -266,7 +267,7 @@ class TestAtoms(BaseTest):
         """
         atom = matrix_frac(self.x, self.A)
         self.assertEquals(atom.size, (1,1))
-        self.assertEquals(atom.curvature, u.Curvature.CONVEX_KEY)
+        self.assertEquals(atom.curvature, s.CONVEX)
         # Test matrix_frac size validation.
         with self.assertRaises(Exception) as cm:
             matrix_frac(self.x, self.C)
@@ -282,10 +283,10 @@ class TestAtoms(BaseTest):
         """Test max_entries.
         """
         # One arg, test sign.
-        self.assertEquals(max_entries(1).sign, u.Sign.POSITIVE_KEY)
-        self.assertEquals(max_entries(-2).sign, u.Sign.NEGATIVE_KEY)
-        self.assertEquals(max_entries(Variable()).sign, u.Sign.UNKNOWN_KEY)
-        self.assertEquals(max_entries(0).sign, u.Sign.ZERO_KEY)
+        self.assertEquals(max_entries(1).sign, s.POSITIVE)
+        self.assertEquals(max_entries(-2).sign, s.NEGATIVE)
+        self.assertEquals(max_entries(Variable()).sign, s.UNKNOWN)
+        self.assertEquals(max_entries(0).sign, s.ZERO)
 
         # Test with axis argument.
         self.assertEquals(max_entries(Variable(2), axis=0).size, (1, 1))
@@ -304,10 +305,10 @@ class TestAtoms(BaseTest):
         """Test min_entries.
         """
         # One arg, test sign.
-        self.assertEquals(min_entries(1).sign, u.Sign.POSITIVE_KEY)
-        self.assertEquals(min_entries(-2).sign, u.Sign.NEGATIVE_KEY)
-        self.assertEquals(min_entries(Variable()).sign, u.Sign.UNKNOWN_KEY)
-        self.assertEquals(min_entries(0).sign, u.Sign.ZERO_KEY)
+        self.assertEquals(min_entries(1).sign, s.POSITIVE)
+        self.assertEquals(min_entries(-2).sign, s.NEGATIVE)
+        self.assertEquals(min_entries(Variable()).sign, s.UNKNOWN)
+        self.assertEquals(min_entries(0).sign, s.ZERO)
 
         # Test with axis argument.
         self.assertEquals(min_entries(Variable(2), axis=0).size, (1, 1))
@@ -325,69 +326,69 @@ class TestAtoms(BaseTest):
     # Test sign logic for max_elemwise.
     def test_max_elemwise_sign(self):
         # Two args.
-        self.assertEquals(max_elemwise(1, 2).sign, u.Sign.POSITIVE_KEY)
-        self.assertEquals(max_elemwise(1, Variable()).sign, u.Sign.POSITIVE_KEY)
-        self.assertEquals(max_elemwise(1, -2).sign, u.Sign.POSITIVE_KEY)
-        self.assertEquals(max_elemwise(1, 0).sign, u.Sign.POSITIVE_KEY)
+        self.assertEquals(max_elemwise(1, 2).sign, s.POSITIVE)
+        self.assertEquals(max_elemwise(1, Variable()).sign, s.POSITIVE)
+        self.assertEquals(max_elemwise(1, -2).sign, s.POSITIVE)
+        self.assertEquals(max_elemwise(1, 0).sign, s.POSITIVE)
 
-        self.assertEquals(max_elemwise(Variable(), 0).sign, u.Sign.POSITIVE_KEY)
-        self.assertEquals(max_elemwise(Variable(), Variable()).sign, u.Sign.UNKNOWN_KEY)
-        self.assertEquals(max_elemwise(Variable(), -2).sign, u.Sign.UNKNOWN_KEY)
+        self.assertEquals(max_elemwise(Variable(), 0).sign, s.POSITIVE)
+        self.assertEquals(max_elemwise(Variable(), Variable()).sign, s.UNKNOWN)
+        self.assertEquals(max_elemwise(Variable(), -2).sign, s.UNKNOWN)
 
-        self.assertEquals(max_elemwise(0, 0).sign, u.Sign.ZERO_KEY)
-        self.assertEquals(max_elemwise(0, -2).sign, u.Sign.ZERO_KEY)
+        self.assertEquals(max_elemwise(0, 0).sign, s.ZERO)
+        self.assertEquals(max_elemwise(0, -2).sign, s.ZERO)
 
-        self.assertEquals(max_elemwise(-3, -2).sign, u.Sign.NEGATIVE_KEY)
+        self.assertEquals(max_elemwise(-3, -2).sign, s.NEGATIVE)
 
         # Many args.
         self.assertEquals(max_elemwise(-2, Variable(), 0, -1, Variable(), 1).sign,
-                          u.Sign.POSITIVE_KEY)
+                          s.POSITIVE)
 
         # Promotion.
         self.assertEquals(max_elemwise(1, Variable(2)).sign,
-                          u.Sign.POSITIVE_KEY)
+                          s.POSITIVE)
         self.assertEquals(max_elemwise(1, Variable(2)).size,
                           (2, 1))
 
     # Test sign logic for min_elemwise.
     def test_min_elemwise_sign(self):
         # Two args.
-        self.assertEquals(min_elemwise(1, 2).sign, u.Sign.POSITIVE_KEY)
-        self.assertEquals(min_elemwise(1, Variable()).sign, u.Sign.UNKNOWN_KEY)
-        self.assertEquals(min_elemwise(1, -2).sign, u.Sign.NEGATIVE_KEY)
-        self.assertEquals(min_elemwise(1, 0).sign, u.Sign.ZERO_KEY)
+        self.assertEquals(min_elemwise(1, 2).sign, s.POSITIVE)
+        self.assertEquals(min_elemwise(1, Variable()).sign, s.UNKNOWN)
+        self.assertEquals(min_elemwise(1, -2).sign, s.NEGATIVE)
+        self.assertEquals(min_elemwise(1, 0).sign, s.ZERO)
 
-        self.assertEquals(min_elemwise(Variable(), 0).sign, u.Sign.NEGATIVE_KEY)
-        self.assertEquals(min_elemwise(Variable(), Variable()).sign, u.Sign.UNKNOWN_KEY)
-        self.assertEquals(min_elemwise(Variable(), -2).sign, u.Sign.NEGATIVE_KEY)
+        self.assertEquals(min_elemwise(Variable(), 0).sign, s.NEGATIVE)
+        self.assertEquals(min_elemwise(Variable(), Variable()).sign, s.UNKNOWN)
+        self.assertEquals(min_elemwise(Variable(), -2).sign, s.NEGATIVE)
 
-        self.assertEquals(min_elemwise(0, 0).sign, u.Sign.ZERO_KEY)
-        self.assertEquals(min_elemwise(0, -2).sign, u.Sign.NEGATIVE_KEY)
+        self.assertEquals(min_elemwise(0, 0).sign, s.ZERO)
+        self.assertEquals(min_elemwise(0, -2).sign, s.NEGATIVE)
 
-        self.assertEquals(min_elemwise(-3, -2).sign, u.Sign.NEGATIVE_KEY)
+        self.assertEquals(min_elemwise(-3, -2).sign, s.NEGATIVE)
 
         # Many args.
         self.assertEquals(min_elemwise(-2, Variable(), 0, -1, Variable(), 1).sign,
-                          u.Sign.NEGATIVE_KEY)
+                          s.NEGATIVE)
 
         # Promotion.
         self.assertEquals(min_elemwise(-1, Variable(2)).sign,
-                          u.Sign.NEGATIVE_KEY)
+                          s.NEGATIVE)
         self.assertEquals(min_elemwise(-1, Variable(2)).size,
                           (2, 1))
 
     def test_sum_entries(self):
         """Test the sum_entries atom.
         """
-        self.assertEquals(sum_entries(1).sign, u.Sign.POSITIVE_KEY)
-        self.assertEquals(sum_entries([1, -1]).sign, u.Sign.UNKNOWN_KEY)
-        self.assertEquals(sum_entries([1, -1]).curvature, u.Curvature.CONSTANT_KEY)
-        self.assertEquals(sum_entries(Variable(2)).sign, u.Sign.UNKNOWN_KEY)
+        self.assertEquals(sum_entries(1).sign, s.POSITIVE)
+        self.assertEquals(sum_entries([1, -1]).sign, s.UNKNOWN)
+        self.assertEquals(sum_entries([1, -1]).curvature, s.CONSTANT)
+        self.assertEquals(sum_entries(Variable(2)).sign, s.UNKNOWN)
         self.assertEquals(sum_entries(Variable(2)).size, (1, 1))
-        self.assertEquals(sum_entries(Variable(2)).curvature, u.Curvature.AFFINE_KEY)
+        self.assertEquals(sum_entries(Variable(2)).curvature, s.AFFINE)
         # Mixed curvature.
         mat = np.mat("1 -1")
-        self.assertEquals(sum_entries(mat*square(Variable(2))).curvature, u.Curvature.UNKNOWN_KEY)
+        self.assertEquals(sum_entries(mat*square(Variable(2))).curvature, s.UNKNOWN)
 
         # Test with axis argument.
         self.assertEquals(sum_entries(Variable(2), axis=0).size, (1, 1))
@@ -405,16 +406,16 @@ class TestAtoms(BaseTest):
     def test_mul_elemwise(self):
         """Test the mul_elemwise atom.
         """
-        self.assertEquals(mul_elemwise([1, -1], self.x).sign, u.Sign.UNKNOWN_KEY)
-        self.assertEquals(mul_elemwise([1, -1], self.x).curvature, u.Curvature.AFFINE_KEY)
+        self.assertEquals(mul_elemwise([1, -1], self.x).sign, s.UNKNOWN)
+        self.assertEquals(mul_elemwise([1, -1], self.x).curvature, s.AFFINE)
         self.assertEquals(mul_elemwise([1, -1], self.x).size, (2, 1))
         pos_param = Parameter(2, sign="positive")
         neg_param = Parameter(2, sign="negative")
-        self.assertEquals(mul_elemwise(pos_param, pos_param).sign, u.Sign.POSITIVE_KEY)
-        self.assertEquals(mul_elemwise(pos_param, neg_param).sign, u.Sign.NEGATIVE_KEY)
-        self.assertEquals(mul_elemwise(neg_param, neg_param).sign, u.Sign.POSITIVE_KEY)
+        self.assertEquals(mul_elemwise(pos_param, pos_param).sign, s.POSITIVE)
+        self.assertEquals(mul_elemwise(pos_param, neg_param).sign, s.NEGATIVE)
+        self.assertEquals(mul_elemwise(neg_param, neg_param).sign, s.POSITIVE)
 
-        self.assertEquals(mul_elemwise(neg_param, square(self.x)).curvature, u.Curvature.CONCAVE_KEY)
+        self.assertEquals(mul_elemwise(neg_param, square(self.x)).curvature, s.CONCAVE)
 
         # Test promotion.
         self.assertEquals(mul_elemwise([1, -1], 1).size, (2, 1))
@@ -456,16 +457,16 @@ class TestAtoms(BaseTest):
         """Test the reshape class.
         """
         expr = reshape(self.A, 4, 1)
-        self.assertEquals(expr.sign, u.Sign.UNKNOWN_KEY)
-        self.assertEquals(expr.curvature, u.Curvature.AFFINE_KEY)
+        self.assertEquals(expr.sign, s.UNKNOWN)
+        self.assertEquals(expr.curvature, s.AFFINE)
         self.assertEquals(expr.size, (4, 1))
 
         expr = reshape(expr, 2, 2)
         self.assertEquals(expr.size, (2, 2))
 
         expr = reshape(square(self.x), 1, 2)
-        self.assertEquals(expr.sign, u.Sign.POSITIVE_KEY)
-        self.assertEquals(expr.curvature, u.Curvature.CONVEX_KEY)
+        self.assertEquals(expr.sign, s.POSITIVE)
+        self.assertEquals(expr.curvature, s.CONVEX)
         self.assertEquals(expr.size, (1, 2))
 
         with self.assertRaises(Exception) as cm:
@@ -477,34 +478,34 @@ class TestAtoms(BaseTest):
         """Test the vec atom.
         """
         expr = vec(self.C)
-        self.assertEquals(expr.sign, u.Sign.UNKNOWN_KEY)
-        self.assertEquals(expr.curvature, u.Curvature.AFFINE_KEY)
+        self.assertEquals(expr.sign, s.UNKNOWN)
+        self.assertEquals(expr.curvature, s.AFFINE)
         self.assertEquals(expr.size, (6, 1))
 
         expr = vec(self.x)
         self.assertEquals(expr.size, (2, 1))
 
         expr = vec(square(self.a))
-        self.assertEquals(expr.sign, u.Sign.POSITIVE_KEY)
-        self.assertEquals(expr.curvature, u.Curvature.CONVEX_KEY)
+        self.assertEquals(expr.sign, s.POSITIVE)
+        self.assertEquals(expr.curvature, s.CONVEX)
         self.assertEquals(expr.size, (1, 1))
 
     def test_diag(self):
         """Test the diag atom.
         """
         expr = diag(self.x)
-        self.assertEquals(expr.sign, u.Sign.UNKNOWN_KEY)
-        self.assertEquals(expr.curvature, u.Curvature.AFFINE_KEY)
+        self.assertEquals(expr.sign, s.UNKNOWN)
+        self.assertEquals(expr.curvature, s.AFFINE)
         self.assertEquals(expr.size, (2, 2))
 
         expr = diag(self.A)
-        self.assertEquals(expr.sign, u.Sign.UNKNOWN_KEY)
-        self.assertEquals(expr.curvature, u.Curvature.AFFINE_KEY)
+        self.assertEquals(expr.sign, s.UNKNOWN)
+        self.assertEquals(expr.curvature, s.AFFINE)
         self.assertEquals(expr.size, (2, 1))
 
         expr = diag(self.x.T)
-        self.assertEquals(expr.sign, u.Sign.UNKNOWN_KEY)
-        self.assertEquals(expr.curvature, u.Curvature.AFFINE_KEY)
+        self.assertEquals(expr.sign, s.UNKNOWN)
+        self.assertEquals(expr.curvature, s.AFFINE)
         self.assertEquals(expr.size, (2, 2))
 
         with self.assertRaises(Exception) as cm:
@@ -516,8 +517,8 @@ class TestAtoms(BaseTest):
         """Test the trace atom.
         """
         expr = trace(self.A)
-        self.assertEquals(expr.sign, u.Sign.UNKNOWN_KEY)
-        self.assertEquals(expr.curvature, u.Curvature.AFFINE_KEY)
+        self.assertEquals(expr.sign, s.UNKNOWN)
+        self.assertEquals(expr.curvature, s.AFFINE)
         self.assertEquals(expr.size, (1, 1))
 
         with self.assertRaises(Exception) as cm:
@@ -529,11 +530,11 @@ class TestAtoms(BaseTest):
         """Test the log1p atom.
         """
         expr = log1p(1)
-        self.assertEquals(expr.sign, u.Sign.POSITIVE_KEY)
-        self.assertEquals(expr.curvature, u.Curvature.CONSTANT_KEY)
+        self.assertEquals(expr.sign, s.POSITIVE)
+        self.assertEquals(expr.curvature, s.CONSTANT)
         self.assertEquals(expr.size, (1, 1))
         expr = log1p(-0.5)
-        self.assertEquals(expr.sign, u.Sign.NEGATIVE_KEY)
+        self.assertEquals(expr.sign, s.NEGATIVE)
 
     def test_upper_tri(self):
         with self.assertRaises(Exception) as cm:
@@ -710,11 +711,11 @@ class TestAtoms(BaseTest):
         xval = [-5]*dims
         p2 = Problem(cvxpy.Minimize(sum_entries(t)), [-t<=x, x<=t])
         g = cvxpy.partial_optimize(p2, [t], [x])
-        self.assertEquals(g.curvature, u.Curvature.CONVEX_KEY)
+        self.assertEquals(g.curvature, s.CONVEX)
 
         p2 = Problem(cvxpy.Maximize(sum_entries(t)), [-t<=x, x<=t])
         g = cvxpy.partial_optimize(p2, [t], [x])
-        self.assertEquals(g.curvature, u.Curvature.CONCAVE_KEY)
+        self.assertEquals(g.curvature, s.CONCAVE)
 
     # Test the partial_optimize atom.
     def test_partial_optimize_eval_1norm(self):