+++ CHANGELOG.md
@@ -23,9 +23,7 @@ Version 0.3.10 (next release)
 ----------------------------
 * TODO make division work elementwise (and with division of scalar by variable)
 * TODO test and fix updating constraints.
-* TODO make curvature, sign recursive and affine == convex + concave || constant.
-Will fix bug where change constant matrix after the fact.
-Constant is separate (check if no variables).
+* Made curvature, sign recursive and affine == convex + concave || constant.
 
 Version 0.3.9
 -------------